Estimate Posthoc

Provides a table of confidence intervals for all Tukey HSD comparisons.

Usage

# Default S3 method
estimatePosthoc(frame, conf.level=.95, main=NULL, digits=3)

# S3 method for class 'formula'
estimatePosthoc(formula, conf.level=.95, main=NULL, digits=3)

# S3 method for class 'wss'
estimatePosthoc(DescStats, conf.level=.95, main=NULL, digits=3)

# S3 method for class 'bss'
estimatePosthoc(DescStats, conf.level=.95, main=NULL, digits=3)

Arguments

frame

a set of (non-empty) numeric vectors of data values

formula

a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs a factor giving the corresponding groups

DescStats

a matrix/frame of descriptive/summary statistics

conf.level

confidence level of the interval

main

a custom title for the table

digits

the specified number of decimal places

Value

The returned frame contains the following components:

Diff

the differences between the pairs of means

SE

the standard error of the differences

df

the degrees of the freedom for the differences

LL

the lower limits of the confidence intervals for the differences

UL

the upper limits of the confidence intervals for the differences