Standardize Contrast

Provides the standardized effect size and its confidence interval for the contrast.

Usage

# S3 method for class 'data.frame'
standardizeContrast(frame, contrast, mu=0, conf.level=.95, ...)

# S3 method for class 'formula'
standardizeContrast(formula, contrast, mu=0, conf.level=.95, ...)

# S3 method for class 'wsm'
standardizeContrast(moments, corrs, contrast, mu=0, conf.level=.95, ...)

# S3 method for class 'bsm'
standardizeContrast(moments, contrast, mu=0, conf.level=.95, ...)

Arguments

frame

a set of (non-empty) numeric vectors of data values

formula

a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs a factor giving the corresponding groups

moments

a matrix/frame of descriptive/summary statistics

corrs

a matrix/frame of correlations among the measures

contrast

a vector of codes for the contrast among all means

mu

a number indicating the true/hypothesized value of the mean contrast

conf.level

confidence level of the interval

Value

The returned frame contains the following components:

d

the standardized mean contrast

SE

the standard errors of the standardized mean contrast

LL

the lower limit of the confidence interval of the standardized mean contrast

UL

the upper limit of the confidence interval of the standardized mean contrast

References

Bonett, D. G. (2008). Confidence intervals for standardized linear contrasts of means. Psychological Methods, 13(2), 99-109.